requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.1
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
openai==0.27.5
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.2
yarl==1.9.2
